Release v0.0.20: Streamline settings UI architecture
- Create FretboardSettings component for consolidated fretboard display settings
- Move tuning, fret count, invert, and inlays to inline bar with fretboard
- Move Intervals toggle and Filter dropdown to FretboardSettings (Learn mode)
- Replace GlobalSettings with simplified JamSettings (2 sections vs 4 collapsible)
- Add "Vol:" prefix to volume sliders in Jam settings
- Restore BPM slider in transport controls
- Simplify SequenceStep to only time signature override
- Remove duplicate settings controls across modes
