--- src/components/Jam/GlobalSettings.css
@@ -1,413 +1,0 @@
-/* Global Settings Container */
-.global-settings {
-  background: var(--bg-secondary);
-  border-radius: 8px;
-  margin-bottom: 20px;
-  overflow: hidden;
-}
-
-.global-settings-toggle {
-  width: 100%;
-  padding: 15px 20px;
-  background: var(--bg-secondary);
-  border: none;
-  color: var(--text-primary);
-  font-size: 1em;
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  cursor: pointer;
-  transition: background 0.2s ease;
-}
-
-.global-settings-toggle:hover {
-  background: var(--bg-tertiary);
-}
-
-.toggle-icon {
-  color: var(--accent);
-  font-size: 0.8em;
-  transition: transform 0.2s ease;
-}
-
-.settings-badge {
-  background: var(--accent);
-  color: #fff;
-  font-size: 0.75em;
-  padding: 2px 8px;
-  border-radius: 10px;
-  margin-left: auto;
-}
-
-.global-settings-content {
-  padding: 20px;
-  border-top: 1px solid var(--bg-tertiary);
-  animation: slideDown 0.2s ease;
-}
-
-@keyframes slideDown {
-  from {
-    opacity: 0;
-    transform: translateY(-10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-/* Settings Sections */
-.settings-section {
-  margin-bottom: 25px;
-  padding-bottom: 20px;
-  border-bottom: 1px solid var(--bg-tertiary);
-}
-
-.settings-section:last-of-type {
-  border-bottom: none;
-  margin-bottom: 15px;
-}
-
-.section-header {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-  user-select: none;
-  padding: 6px 8px;
-  margin: 0 -8px 12px -8px;
-  border-radius: 4px;
-  transition: background 0.2s ease;
-}
-
-.section-header:hover {
-  background: var(--bg-primary);
-}
-
-.section-chevron {
-  color: var(--accent);
-  font-size: 0.7em;
-  transition: transform 0.2s ease;
-}
-
-.settings-section h4 {
-  margin: 0;
-  color: var(--text-primary);
-  font-size: 1em;
-  flex: 1;
-}
-
-/* Time Signature Controls */
-.time-sig-controls,
-.accent-controls,
-.metronome-controls,
-.chord-audio-controls {
-  animation: slideDown 0.2s ease;
-}
-
-.time-sig-controls {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.preset-buttons {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-}
-
-.preset-btn {
-  padding: 6px 14px;
-  background: var(--bg-primary);
-  border: 1px solid var(--bg-tertiary);
-  color: var(--text-primary);
-  border-radius: 4px;
-  font-size: 0.9em;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.preset-btn:hover {
-  background: var(--bg-tertiary);
-  border-color: var(--accent);
-}
-
-.preset-btn.active {
-  background: var(--accent);
-  border-color: var(--accent);
-  color: #fff;
-}
-
-.custom-time-sig {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.custom-time-sig label {
-  color: var(--text-secondary);
-  font-size: 0.9em;
-}
-
-.time-sig-input {
-  width: 50px;
-  padding: 6px;
-  background: var(--bg-primary);
-  border: 1px solid var(--bg-tertiary);
-  border-radius: 4px;
-  color: var(--text-primary);
-  text-align: center;
-  font-size: 0.95em;
-}
-
-.time-sig-select {
-  padding: 6px 8px;
-  background: var(--bg-primary);
-  border: 1px solid var(--bg-tertiary);
-  border-radius: 4px;
-  color: var(--text-primary);
-  font-size: 0.95em;
-}
-
-/* Accent Pattern Controls */
-.accent-controls {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.radio-group {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.radio-label {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  color: var(--text-secondary);
-  cursor: pointer;
-  padding: 6px 8px;
-  border-radius: 4px;
-  transition: background 0.2s ease;
-}
-
-.radio-label:hover {
-  background: var(--bg-primary);
-}
-
-.radio-label input[type="radio"] {
-  accent-color: var(--accent);
-  cursor: pointer;
-}
-
-.radio-label input[type="radio"]:checked + span {
-  color: var(--text-primary);
-  font-weight: 500;
-}
-
-.custom-accents {
-  padding: 12px;
-  background: var(--bg-primary);
-  border-radius: 6px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.custom-accents label {
-  color: var(--text-secondary);
-  font-size: 0.9em;
-}
-
-.accent-toggles {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-}
-
-.accent-toggle-btn {
-  width: 40px;
-  height: 40px;
-  background: var(--bg-secondary);
-  border: 2px solid var(--bg-tertiary);
-  color: var(--text-secondary);
-  border-radius: 6px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.accent-toggle-btn:hover {
-  border-color: var(--accent);
-}
-
-.accent-toggle-btn.active {
-  background: var(--accent);
-  border-color: var(--accent);
-  color: #fff;
-  box-shadow: 0 0 8px rgba(233, 69, 96, 0.3);
-}
-
-.swing-controls {
-  padding: 12px;
-  background: var(--bg-primary);
-  border-radius: 6px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.swing-controls label {
-  color: var(--text-primary);
-  font-size: 0.9em;
-  font-weight: 500;
-}
-
-.swing-slider {
-  width: 100%;
-  accent-color: var(--accent);
-  cursor: pointer;
-}
-
-.swing-labels {
-  display: flex;
-  justify-content: space-between;
-  color: var(--text-secondary);
-  font-size: 0.85em;
-}
-
-/* Metronome Sound Controls */
-.metronome-controls {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.control-row {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.control-row label {
-  color: var(--text-secondary);
-  font-size: 0.9em;
-  min-width: 100px;
-}
-
-.sound-select,
-.subdivision-select {
-  flex: 1;
-  padding: 8px 10px;
-  background: var(--bg-primary);
-  border: 1px solid var(--bg-tertiary);
-  border-radius: 4px;
-  color: var(--text-primary);
-  font-size: 0.95em;
-}
-
-/* Chord Audio Controls */
-.chord-audio-controls {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.volume-slider,
-.duration-slider {
-  width: 100%;
-  accent-color: var(--accent);
-  cursor: pointer;
-}
-
-.waveform-select {
-  flex: 1;
-  padding: 8px 10px;
-  background: var(--bg-primary);
-  border: 1px solid var(--bg-tertiary);
-  border-radius: 4px;
-  color: var(--text-primary);
-  font-size: 0.95em;
-}
-
-.toggle-label {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  color: var(--text-secondary);
-  cursor: pointer;
-  padding: 6px 8px;
-  border-radius: 4px;
-  transition: all 0.2s ease;
-}
-
-.toggle-label:hover {
-  background: var(--bg-primary);
-}
-
-.toggle-label.active {
-  color: var(--text-primary);
-  font-weight: 500;
-}
-
-.toggle-label input[type="checkbox"] {
-  accent-color: var(--accent);
-  cursor: pointer;
-  width: 18px;
-  height: 18px;
-}
-
-/* Settings Actions */
-.settings-actions {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 10px;
-}
-
-.reset-btn {
-  padding: 8px 16px;
-  background: var(--bg-tertiary);
-  border: 1px solid var(--bg-tertiary);
-  color: var(--text-secondary);
-  border-radius: 4px;
-  font-size: 0.9em;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.reset-btn:hover {
-  background: var(--bg-primary);
-  color: var(--accent);
-  border-color: var(--accent);
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-  .preset-buttons {
-    gap: 6px;
-  }
-
-  .preset-btn {
-    padding: 5px 10px;
-    font-size: 0.85em;
-  }
-
-  .control-row {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .control-row label {
-    min-width: auto;
-  }
-
-  .sound-select,
-  .subdivision-select {
-    width: 100%;
-  }
-}
